Move FilterCollections into components folder
