+++ src/pages/LiquidationsPage/components/RafflesList/RafflesList.tsx
@@ -2,7 +2,7 @@ import { FC, ReactNode, useRef } from 'react';
 import { Controller } from 'react-hook-form';
 
 import { FilterFormInputsNames, RafflesListFormNames } from '../../model';
-import FilterCollections from '@frakt/componentsNew/FilterCollections';
+import FilterCollections from '@frakt/components/FilterCollections';
 import SortControl from '@frakt/componentsNew/SortControl';
 import { TicketsCounter } from '../TicketsCounter';
 import { useOnClickOutside } from '@frakt/hooks';
